Replace the SoE function entirely, add comments
Instead of putting found primes to a given stream, the function now modifies a bitset of custom length passed to it. The principles stay almost the same, but the burden to use the result numbers is now on the client.
This allows to only check for primality numbers within [2; sqrt(N)] range, which fastens the function significantly.

The function is now a template and only takes a bitset of a custom uint64_t size.

Also, better descriptions of the algorithm and the function were provided.