Stop methods from returning references to pointers
Inner methods looking for certain nodes do not attempt to return
references to pointers anymore. The aforementioned methods are marked as
const now since they do not make any modifications.

The idea behind returning references to pointers was to be able to use
the methods as lvalues. However, there is no actual need to do that; you
only need to access members of Node structures which can be done via
pointers as they are.